requirements.txt
fastapi
uvicorn
openai
sentence-transformers
newspaper3k
duckduckgo-search
faiss-cpu
unstructured
langchain
langchain-unstructured
langchain-community
langchain-huggingface
psycopg2-binary
beautifulsoup4
lxml
python-dotenv